requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
clarabel==0.11.1
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
curl-cffi==0.13.0
cvxpy==1.7.5
cycler==0.12.1
et-xmlfile==2.0.0
exceptiongroup==1.3.1
fonttools==4.62.1
frozendict==2.4.7
gitdb==4.0.12
gitpython==3.1.46
h11==0.16.0
hmmlearn==0.3.3
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lxml==6.0.2
markupsafe==3.0.3
matplotlib==3.10.8
multitasking==0.0.12
mypy-extensions==1.1.0
narwhals==2.18.1
numpy==2.2.6
openpyxl==3.1.5
osqp==1.1.1
packaging==26.0
pandas==2.3.3
pandera==0.30.1
patsy==1.0.2
peewee==4.0.2
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
protobuf==6.33.6
pyarrow==23.0.1
pycoingecko==3.2.0
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
pyportfolioopt==1.6.0
python-bcb==0.3.4
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-base==0.13.1
scikit-learn==1.7.2
scipy==1.15.3
scs==3.2.11
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
statsmodels==0.14.6
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typeguard==4.5.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
websockets==16.0
yfinance==1.2.0
